Don't swim against the tide…
Initially, the `onChange` and `onSelect` props to be are not
invoked on the initial render since they aren’t required so
`useDidMountEffect` was created. But this is “swimming against the
tide” (see link) and goes React's *against* React's fundamentals.
This commit allows React to synchronize all state and props, including
initial state and props.

The commit also addresses a few other concerns:

1. Memoizes the `onSelect` and `onChange` props passed into the
uncontrolled example. Non memoized props cause unnecessary renders of
 the Listbox.

2. Only call `onSelect` and `onChange` props in `useEffect` for
uncontrolled components. This is because for controlled components,
the prop callbacks are invoked in their event handler functions

See: https://overreacted.io/a-complete-guide-to-useeffect/#swimming-against-the-tide