Implement grouped listbox
The significant change in this PR involves moving population of
`options` mutableRef out of `getOptionProps` because `getOptionProps`
can possibiliy render more than once due to Concurrent React. Because
of this, we could get an inaccurate index for each option.

The logic which populates the `options` ref has been moved into a
`useEffect` at the component level.