Make indices stable and concurrent safe
Turns out the indices generated are not reliable for React Concurrent
mode. The index ref is incremented within a useRef, which gets called
twice in `React.StrictMode`. This causes the indices to increment to an
unstable, inaccurate index.

This commit wraps a Storybook example in `React.StrictMode` and adds a
custom hook to compute the index, and triggers an additional render
with the found index. This does cause an extra render, but shouldn't be
too expensive since it's a single ListboxOption component.

Now, when `React.Strict` mode triggers double renders, the index
remains stable.